requirements.txt
fastapi
mysql-connector-python
SQLAlchemy
pymysql